pyproject.toml
[tool.poetry]
name = "repl_nix_VIP-3"
version = "0.1.0"
description = ""
authors = ["Replit user <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
discum = "^1.4.1"
inputimeout = "^1.0.4"
requests = "^2.28.1"
TwoCaptcha = "^0.0.1"
colorama = "^0.4.6"
discord-webhook = "^1.0.0"
Flask = "^2.2.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
